+++ deepmatch/utils.py
@@ -36,7 +36,7 @@ class NegativeSampler(
     __slots__ = ()
 
     def __new__(cls, sampler, num_sampled, item_name, item_count=None, distortion=1.0, ):
-        if sampler not in ['in_batch', 'uniform', 'fixed_unigram', 'learned_unigram']:
+        if sampler not in ['inbatch', 'uniform', 'frequency', 'adaptive']:
             raise ValueError(' `%s` sampler is not supported ' % sampler)
         return super(NegativeSampler, cls).__new__(cls, sampler, num_sampled, item_name, item_count, distortion)